Move members.ts to routers folder
